+++ tests/Sample/Handlers/ToJson.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Rougin\Slytherin\Sample\Handlers;
+
+use Interop\Http\ServerMiddleware\DelegateInterface;
+use Interop\Http\ServerMiddleware\MiddlewareInterface;
+use Psr\Http\Message\ServerRequestInterface;
+
+/**
+ * @package Slytherin
+ * @author  Rougin Gutib <rougingutib@gmail.com>
+ */
+class ToJson implements MiddlewareInterface
+{
+    public function process(ServerRequestInterface $request, DelegateInterface $delegate)
+    {
+        $response = $delegate->process($request);
+
+        return $response->withHeader('Content-Type', 'application/json');
+    }
+}